### 2025-10-06 #9153 @editorialbot
[PRE REVIEW]: A finite element environment for dynamic simulation of beam networks and lattice metamaterials
**Submitting author:** @hortulanusT (<a href="http://orcid.org/0000-0002-5088-855X">Til Gärtner</a>)
**Repository:** https://github.com/hortulanusT/dynLattice
**Branch with paper.md** (empty if default branch):
**Version:** 1.11
