### 2025-10-08 45ee65a
added relevant paper from JOSS pre-review
