+++ paper/paper.bib
@@ -119,7 +119,6 @@
   day      = {03},
   doi      = {10.4233/uuid:9d0b230e-1008-46a2-a193-b9dd3f8c472f},
   month    = {11},
-  note     = {to be published},
   school   = {Technische Universiteit Delft},
   subtitle = {On the Behaviour of Architected Metamaterials under High-Rate Loading Conditions},
   title    = {Are Auxetics Better for Protection?},
@@ -259,3 +258,16 @@
   volume  = {122},
   year    = {2021}
 }
+
+@article{GXBeam,
+  doi       = {10.21105/joss.03997},
+  url       = {https://doi.org/10.21105/joss.03997},
+  year      = {2022},
+  publisher = {The Open Journal},
+  volume    = {7},
+  number    = {73},
+  pages     = {3997},
+  author    = {McDonnell, Taylor and Ning, Andrew},
+  title     = {GXBeam: A Pure Julia Implementation of Geometrically Exact Beam Theory},
+  journal   = {Journal of Open Source Software}
+} 
