### 2025-10-08 #9153 @hortulanusT
Hello @kyleniemeyer thank you for your processing of our submission!
I went through the comments of the editorialbot and fixed us missing out on the GXBeam paper (https://github.com/hortulanusT/dynLattice/pull/7).
From the other comments I do not think there is something for me to do, I only want to emphasize, that the suggested DOI for the JIVE entry is wrong, there does not exist a DOI for the JIVE package, so I believe it should be sorted under the "SKIP DOI" header, if there is anything I can do to facilitate this, please let me know!

Regarding the reviewers I would suggest the following 5 from the database you linked:
- thelfer
- Onur Deniz Akan (odakan)
- Konstantinos Mixios (iammix)
- chennachaos
- CWillberg

Next to them I would like to suggest (he was the supervisor of my MSc thesis 5 years ago, if this would represent a conflict of interest, please disregard this suggestion)
- Mauricio Fernández (mauricio-fernandez-l)
