### 2026-01-21 #9153 @hortulanusT
@kyleniemeyer @mbarzegary Thank you for the update!

I will have to read a bit into the new Scope and also discuss the changed requirements with my co-authors. I will of course let you know as soon as I have more clarity on my side how to proceed.
