### 2026-01-26 414461b
changes according to https://github.com/openjournals/joss-reviews/issues/9153#issuecomment-3773737129
