+++ paper/paper.md
@@ -9,7 +9,7 @@ tags:
 authors:
   - name: Til Gärtner
     orcid: 0000-0002-5088-855X
-    affiliation: "1, 2"
+    affiliation: "1, 2, 3"
   - name: Frans P. van der Meer
     orcid: 0000-0002-6691-1259
     affiliation: 1
@@ -36,14 +36,6 @@ In order to enable engineers to design mechanical metamaterials, the relation be
 For large deformations at high rates, this usually requires destructive testing, limiting the speed of developments.
 This limitation can be overcome using efficient numerical tools, allowing for both the accurate and fast description of the inelastic deformation at high rates [@bonfanti24].
 
-# Software Description
-In order to design metamaterials undergoing large deformations at high rates, [dynLattice]{.sc} provides both researchers and engineers in the field a toolkit for numerical experiments.
-[dynLattice]{.sc} is built on top of the openly available C++ `JIVE` finite element toolkit [@jive21; @nguyenthanh20]. 
-This toolkit allows for the implementation of original methods through the development of custom _modules_, for parts of the program flow in the simulation framework, and of custom _models_, for anything related to the numerical representation of the system of partial differential equations.
-In the present addition to the `JIVE` framework, the _models_ further contain a separate _materials_ class, representing the constitutive models considered in the problem.
-This separation allows both the easy and user-friendly extension of the frameworks by simple inheritance of the provided classes as well as usability by structuring the inputs needed for the program flow and the problem description.
-[dynLattice]{.sc} additionally uses `GMSH` [@geuzaine09] to mesh the imported geometries efficiently.
-
 # Statement of need
 There is a wide plethora of existing finite element toolkits available in the market, such as MOOSE [@moose], FEniCS [@fenics], or deal.II [@dealii], to name several prominent ones.
 The usage of these libraries for explicit dynamics with inelastic beams however, would require substantial alteration of their code, as for example in FEniCS explicit dynamics is only available using lumped mass matrices, which would be nonphysical for the rotational inertia contributions of the beams, in deal.II neither beam elements nor explicit solves are available and in MOOSE only a limited Timoshenko-Ehrenfest beam is implemented with only elastic behavior.
@@ -52,6 +44,14 @@ For Kratos similar problems as for MOOSE can be seen in only a limited set of be
 Akantu, focusing on fracture mechanics and providing a solid representation of contact, on the other hand only provides Euler-Bernoulli beam elements, neglecting all shear deformation within the elements.
 Next to general finite element toolkits, there are specialized beam toolkits, most prominently for shear-deformable beams, GXBeam [@GXBeam]. 
 Here, the same beam theory employed in this contribution is used, however due to its focus on composite beams, it is limited to (anisotropic) elastic material behavior and does not model contact.
+
+# Software Design \label{design}
+In order to design metamaterials undergoing large deformations at high rates, [dynLattice]{.sc} provides both researchers and engineers in the field a toolkit for numerical experiments.
+[dynLattice]{.sc} is built on top of the openly available C++ `JIVE` finite element toolkit [@jive21; @nguyenthanh20] and follows in its design the object-oriented framework of `JIVE` to describe the partial differential equations and corresponding solution procedures.
+The toolkit allows for the implementation of original methods through the development of custom _modules_, for parts of the program flow in the simulation framework, and of custom _models_, for anything related to the numerical representation of the system of partial differential equations.
+In the present addition to the `JIVE` framework, the _models_ further contain a separate _materials_ class, representing the constitutive models considered in the problem, sparating the .
[…]
