### 2026-02-06 e0b4e66
Replace autoref with classical link to header
