+++ search/all_17.js
@@ -8,25 +8,26 @@ var searchData=
   ['rdofs_5f_5',['rdofs_',['../classExplicitModule.html#a7c3e31121784bdd72c191a033cfd3546',1,'ExplicitModule']]],
   ['readme_2emd_6',['README.md',['../README_8md.html',1,'']]],
   ['reduction_5fprop_7',['REDUCTION_PROP',['../classNeumannModel.html#ac6a23d6a97151e277c12ce00b1133db0',1,'NeumannModel']]],
-  ['references_8',['References',['..//home/runner/work/dynLattice/dynLattice/paper/paper.md#autotoc_md34',1,'']]],
+  ['references_8',['References',['..//home/runner/work/dynLattice/dynLattice/paper/paper.md#autotoc_md35',1,'']]],
   ['registered_20materials_9',['Registered Materials',['../__declareMaterials_8h.html#Currently',1,'']]],
   ['rejectdeform_10',['rejectdeform',['../classElasticRodMaterial.html#a9700528bd143efdc91c940003aa28704',1,'ElasticRodMaterial::rejectDeform()'],['../classElastoPlasticRodMaterial.html#a9650c029f418f7f20753edb663c5895a',1,'ElastoPlasticRodMaterial::rejectDeform()'],['../classMaterial.html#a750ded2d63b912afa8077758741fefa9',1,'Material::rejectDeform()']]],
   ['related_20publications_11',['Related Publications',['../index.html#autotoc_md24',1,'']]],
   ['report_12',['REPORT',['../testing_8h.html#aa583aec62de195f014c42d5e6ce48491',1,'testing.h']]],
   ['repository_20and_20build_20the_20documentation_13',['2. Get the repository and build the documentation',['../install.html#autotoc_md21',1,'']]],
-  ['requirements_14',['requirements',['../md_README.html#autotoc_md43',1,'Requirements'],['../install.html#autotoc_md18',1,'Requirements']]],
-  ['rod_20constitutive_20models_15',['Materials (Rod Constitutive Models)',['../md_CONTRIBUTING.html#autotoc_md7',1,'']]],
-  ['rod_5fcontact_5fprop_16',['ROD_CONTACT_PROP',['../classLatticeModel.html#a6abc96d8e7dcb6c2ea78281fad051f6a',1,'LatticeModel']]],
-  ['rod_5flist_5fprop_17',['ROD_LIST_PROP',['../classLatticeModel.html#a40a7c6f2b32098ea8ab194db4a2e03cb',1,'LatticeModel']]],
-  ['rodcontactmodel_18',['rodcontactmodel',['../classRodContactModel.html',1,'RodContactModel'],['../classRodContactModel.html#a88c430564bce93755e0749d7722b117f',1,'RodContactModel::RodContactModel()']]],
-  ['rodcontactmodel_2ecpp_19',['RodContactModel.cpp',['../RodContactModel_8cpp.html',1,'']]],
-  ['rodcontactmodel_2eh_20',['RodContactModel.h',['../RodContactModel_8h.html',1,'']]],
-  ['rodname_5f_21',['rodName_',['../classElasticRodMaterial.html#a531a373d5a16daba84a76c9908d8c2e0',1,'ElasticRodMaterial']]],
-  ['rot_5fdof_5fcount_22',['ROT_DOF_COUNT',['../classSpecialCosseratRodModel.html#a1dc97f4994ea59c34ae0b2f71cfcd623',1,'SpecialCosseratRodModel']]],
[…]
