+++ search/all_16.js
@@ -1,5 +1,5 @@
 var searchData=
 [
   ['quadix_0',['Quadix',['../namespacejive__helpers.html#a93805f7b6a667d06c3f465fdce86bd99',1,'jive_helpers']]],
-  ['quick_20start_1',['Quick Start',['../md_README.html#autotoc_md38',1,'']]]
+  ['quick_20start_1',['Quick Start',['../md_README.html#autotoc_md39',1,'']]]
 ];

[…]
