### 2026-02-07 7ef518c
Added State of the field section and ensured the sentence case for all the other headings
