+++ paper/paper.md
@@ -51,8 +51,8 @@ From these measures, strain prescriptors, describing stretching, shearing, bendi
 These strain prescriptors are then used together with the material model of the beam, be it elastic (cf. [@simo86; @eugster15]) or elasto-plastic (cf. [@smriti20; @herrnböck22; @gärtner25]), to assemble the global force vector and---in the static, implicit case---the tangent stiffness matrix using standard finite element procedures.
 
 # Statement of need
-Research on the dynamic behaviour of mechanical metamaterials, especially beam lattices, relies on numerical tools that can capture large deformations, inelastic behaviour, and contact while remaining efficient and flexible enough for parameter studies [@bonfati24].
-General-purpose finite element frameworks, e.g. [@kratos; @Davand2010; @Richart2024], provide broad capabilities but often lack a complete combination of nonlinear Timoshenko-Ehrenfest beam formulations, robust explicit time integration, and inelastic constitutive behavior.
+Research on the dynamic behaviour of mechanical metamaterials, especially beam lattices, relies on numerical tools that can capture large deformations, inelastic behaviour, and contact while remaining efficient and flexible enough for parameter studies [@bonfanti24].
+General-purpose finite element frameworks, e.g. [@kratos; @Dadvand2010; @Richart2024], provide broad capabilities but often lack a complete combination of nonlinear Timoshenko-Ehrenfest beam formulations, robust explicit time integration, and inelastic constitutive behavior.
 Here, a specialized lattice framework, e.g. [@GXBeam], typically delivers high-fidelity beam kinematics but only focuses on elastic deformations or omits contact between the consituting beams of the lattice.
 Resulting from this many researchers rely on their own implementation of Simo-Reissner beams, which are usually not available to the community and only consider a subset of the requirements needed for a general toolkit.
 An open and extensible environment that unifies geometrically exact beams, explicit dynamics, inelasticity, and contact is therefore central to advancing simulation-driven design of lattice metamaterials for high-rate and large deformation applications.
