### 2026-02-07 #9153 @hortulanusT
Hello @kyleniemeyer @mbarzegary,

over the last two weeks @fmeer and I adapted the paper to the new scope of JOSS and would like to move forward with this submission, however, there are two questions, that I am still unsure of:
- The editorial bot keeps mentioning we dont have a `Research impact statement` section, despite one being in both the paper.md and the resulting PDF is there anything we can do for this or is this something that would irrelevant for the further review?
- The DOI entry for JIVE is not missing, the suggested DOI of the editorial bot is misleading, is there some option here to overwrite the suggestion of the bot or is it okay like this?

Thank you again for your kind consideration of this submission!
