### 2026-02-11 #9153 @mbarzegary
@hortulanusT Thanks for updating the paper. Regarding your points:
* The reason is it is spelled as `Resarch` rather than `Research`. Please fix the spelling.
* It's not an issue as long as you say the current DOI is correct.
