### 2026-02-11 #9153 @hortulanusT
@mbarzegary thank you for agreeing to serve as editor for our submission. Thank you as well for pointing out the typo, we could quickly fix.
Regarding the second point: It is correct, that there is no DOI, and I can confirm this.
