### 2026-02-23 #9153 @xuanxu
When cloning the repository git generates this message:
`This repository exceeded its LFS budget. The account responsible for the budget should increase it to restore access.`

The cause seems to be `jive.sif` (size: 1 GB)
