### 2026-02-23 #9153 @hortulanusT
@mbarzegary @xuanxu thanks for pointing this out, I updated the budget for LFS, hoping it works out better now.
We noticed a potential issue if more than 9 people try to clone the repo directly and wrote all documentation in a way it does not collide with my LFS quota (downloading the releases/cloning to the own account for contribution), so I think this might be only conflicting with my quota during review for the moment?
