### 2026-03-13 #9153 @PetrKryslUCSD
I could do this, what is the expected timeframe?

Petr Krysl
4013 Camino Lindo, San Diego, CA 92122

On Thu, Mar 12, 2026, 07:52 Mojtaba Barzegari ***@***.***>
wrote:

> *mbarzegary* left a comment (openjournals/joss-reviews#9153)
> <https://urldefense.com/v3/__https://github.com/openjournals/joss-reviews/issues/9153*issuecomment-4047382999__;Iw!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLRpVqiwg$>
>
> Hi @PetrKryslUCSD
> <https://urldefense.com/v3/__https://github.com/PetrKryslUCSD__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLXBob8HP$>,
> @likask
> <https://urldefense.com/v3/__https://github.com/likask__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLclvudVE$>,
> @jeremylt
> <https://urldefense.com/v3/__https://github.com/jeremylt__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLbj-1hHt$>,
> @YohannDudouit
> <https://urldefense.com/v3/__https://github.com/YohannDudouit__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLWXXVjWA$>,
> @w-luo-academic
> <https://urldefense.com/v3/__https://github.com/w-luo-academic__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLW57VRYG$>,
> and @HaoZeke
> <https://urldefense.com/v3/__https://github.com/HaoZeke__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLXL5UShc$>
> Considering your expertise, would any of you be willing to review this
> submission? As you know, we carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
> <https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/review_criteria.html__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLXpSdX3I$>
>
> If you are not available it would be great if you have any suggestions for
> potential reviewers.
>
> Paper:
> https://github.com/openjournals/joss-papers/blob/joss.09153/joss.09153/10.21105.joss.09153.pdf
> <https://urldefense.com/v3/__https://github.com/openjournals/joss-papers/blob/joss.09153/joss.09153/10.21105.joss.09153.pdf__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLfxw6EA4$>
> Software: https://github.com/hortulanusT/dynLattice
> <https://urldefense.com/v3/__https://github.com/hortulanusT/dynLattice__;!!Mih3wA!BQZZEcLFbu9VyUu3uQo9JornF8aqkFOv7lQdSP_J0B_Q_6RyCQ1e3j7OV1VcXwhN24ImnEyRRflCq42FLfPf-QcN$>
>
