### 2026-03-23 #9153 @PetrKryslUCSD
I'm afraid I will be tied up for the next month. Sorry.

On Mon, Mar 23, 2026, 03:00 Mojtaba Barzegari ***@***.***>
wrote:

> *mbarzegary* left a comment (openjournals/joss-reviews#9153)
> <https://urldefense.com/v3/__https://github.com/openjournals/joss-reviews/issues/9153*issuecomment-4109383970__;Iw!!Mih3wA!H4EnAwqGVgANTBreEWTW5G5k8NG66hs3RHJAQT05AOaOqYol_k6GjKDpizei9YlfVm8hPnkNP2tEUw16IZJj9ZKE$>
>
> @PetrKryslUCSD
> <https://urldefense.com/v3/__https://github.com/PetrKryslUCSD__;!!Mih3wA!H4EnAwqGVgANTBreEWTW5G5k8NG66hs3RHJAQT05AOaOqYol_k6GjKDpizei9YlfVm8hPnkNP2tEUw16IZTHe-Ym$>
> and @jeremylt
> <https://urldefense.com/v3/__https://github.com/jeremylt__;!!Mih3wA!H4EnAwqGVgANTBreEWTW5G5k8NG66hs3RHJAQT05AOaOqYol_k6GjKDpizei9YlfVm8hPnkNP2tEUw16ITUXxejJ$>,
> given the timeframe I mentioned above, would you be willing to review this
> submission?
>
