### 2026-04-01 #9153 @mbarzegary
Hi @thelfer, @SunilAnandatheertha, @TimJosephson, @SoilRos, and @mohd-afeef-badri
Considering your expertise, would any of you be willing to review this submission? As you may know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

If you are not available it would be great if you have any suggestions for potential reviewers.

Paper: https://github.com/openjournals/joss-papers/blob/joss.09153/joss.09153/10.21105.joss.09153.pdf
Software: https://github.com/hortulanusT/dynLattice
