### 2026-04-03 #9153 @SoilRos
Hi @mbarzegary, thanks for the suggestion. I am unfortunately pretty busy at the moment and, honestly, I have been out of the loop of solid mechanics for many years to make a judicious peer-review on this topic. As a potential reviewer I can suggest @sechavarriam. While not currently a JOSS reviewer, I can assert that he would be a much better fit for this than me.

At any rate, I wish you the best with the paper and the software!
