### 2026-04-08 #9153 @sechavarriam
Hi @mbarzegary,

Thank you for reaching out, and thank you @SoilRos for the recommendation.

Yes, I would be willing to review this submission for JOSS. The software topic is indeed within my area of expertise, so I would be glad to contribute to the review process.

Since this would be my first time serving as a reviewer for this journal, I would appreciate it if you could share a clear set of reviewer instructions, including the specific JOSS criteria and guidelines I should keep in mind while evaluating the submission.

Best regards,
@sechavarriam
