### 2026-04-09 #9153 @mbarzegary
Hi @sechavarriam, great, thanks.

This guideline describes the criteria and the objective of the JOSS review process: https://joss.readthedocs.io/en/latest/review_criteria.html. Moreover, the editorial bot will give you a list to go through and complete your review. This is a recent example of how this process looks like: https://github.com/openjournals/joss-reviews/issues/9092, at the beginning of which you can see how the bot generates the checklist and how the reviewers have completed it. In this example, the reviewers have posted the comments in the main thread, although it is recommended to create separate issues (per category) on the software repo to report problems and ping them in the main thread. This is a good example in this regard, with a proper link to the review thread: https://github.com/AlbanBloorRiley/INS_IEP/issues/7. Finally, I will always be there if you have any problems or questions.
