### 2026-05-24 e0dbb3c
Fix OBS-04 of https://github.com/openjournals/joss-reviews/issues/10410#issuecomment-4478050578
