### 2026-05-24 bec079c
fixed typo (mentioned in OBS-09)
