### 2026-05-24 3926c61
fixed shadow declarations (cf OBS-09)
