### 2026-05-24 15d5a62
fixed overloaded-virtual warnings (cf OBS-09)
