### 2026-05-24 47efcf0
fixed old-style-cast warnings (cf OBS-09)
