+++ src/utils/helpers.cpp
@@ -102,7 +102,7 @@ namespace jive_helpers
     double tr_R = trace(R) <= 3. ? trace(R) : 3.; // trace of the matrix
 
     theta = acos((tr_R - 1.) / 2.);
-    rv = unskew((Matrix)(R - R.transpose()));
+    rv = unskew(Matrix(R - R.transpose()));
 
     if (jem::isTiny(theta))
       rv *= 1 / 2.; // infinitesimal rotation
