### 2026-05-24 c1085ea
fixed deprecated-copy warnings (cf OBS-09)
