+++ src/materials/ElastoPlasticRodMaterial.cpp
@@ -63,18 +63,17 @@ void ElastoPlasticRodMaterial::configure(const Properties &props, const Properti
 
   idx_t ipCount;
   idx_t elemCount;
-  idx_t dofCount;
   StringVector dofNames = dofs->getTypeNames();
   String args = StringUtils::join(dofNames, ", ");
 
   myProps.find(ipCount, "ipCount");
   myProps.find(elemCount, "elemCount");
-  dofCount = dofs->typeCount();
-  argCount_ = dofCount;
+  dofCount_ = dofs->typeCount();
+  argCount_ = dofCount_;
 
   double isoCoeff;
   Vector kinHard;
-  Matrix kinFacts(dofCount, dofCount);
+  Matrix kinFacts(dofCount_, dofCount_);
   if (myProps.find(isoCoeff, ISO_HARD_PROP))
   {
     args = args + ", h_0";
@@ -90,10 +89,7 @@ void ElastoPlasticRodMaterial::configure(const Properties &props, const Properti
     }
   }
 
-  stressPart_ = jem::SliceTo(dofCount);
-  hardPart_ = jem::SliceFromTo(dofCount, argCount_);
-
-  materialH_.resize(argCount_ - dofCount, argCount_ - dofCount);
+  materialH_.resize(argCount_ - dofCount_, argCount_ - dofCount_);
   if (argCount_ == 7)
   {
     materialH_ = isoCoeff;
@@ -114,14 +110,14 @@ void ElastoPlasticRodMaterial::configure(const Properties &props, const Properti
         << " ...Hardening matrix of the material '" << myName_ << "':\n"
         << materialH_ << "\n";
 
-  oldHardParams_.resize(argCount_ - dofCount, ipCount, elemCount);
+  oldHardParams_.resize(argCount_ - dofCount_, ipCount, elemCount);
   oldHardParams_ = 0.;
-  currHardParams_.resize(argCount_ - dofCount, ipCount, elemCount);
+  currHardParams_.resize(argCount_ - dofCount_, ipCount, elemCount);
   currHardParams_ = 0.;
 
-  oldPlastStrains_.resize(dofCount, ipCount, elemCount);
+  oldPlastStrains_.resize(dofCount_, ipCount, elemCount);
   oldPlastStrains_ = 0.;
-  currPlastStrains_.resize(dofCount, ipCount, elemCount);
+  currPlastStrains_.resize(dofCount_, ipCount, elemCount);
   currPlastStrains_ = 0.;
 
   currDeltaFlow_.resize(ipCount, elemCount);
@@ -206,7 +202,7 @@ void ElastoPlasticRodMaterial::getStress(const Vector &stress, const Vector &str
   Vector hardParams = oldHardParams_(ALL, ip, ielem).clone();
   double deltaFlow = 0.;
 
-  Vector hardStress(argCount_ - stress.size());
+  Vector hardStress(argCount_ - dofCount_);
   Vector args(argCount_);
   double yieldValue = 0.;
   Vector yieldGrad(argCount_);
@@ -225,8 +221,8 @@ void ElastoPlasticRodMaterial::getStress(const Vector &stress, const Vector &str
       break;
     }
 
-    args[stressPart_] = stress;
-    args[hardPart_] = hardStress;
+    args[jem::SliceTo(dofCount_)] = stress;
+    args[jem::SliceFromTo(dofCount_, argCount_)] = hardStress;
 
     yieldValue = yieldCond_->getValue(args.addr());
 
@@ -248,16 +244,16 @@ void ElastoPlasticRodMaterial::getStress(const Vector &stress, const Vector &str
[…]
