### 2026-05-24 ae3c3fd
fixed suggest-override warnings (cf OBS-09)
