+++ src/models/HingeModel.h
@@ -91,7 +91,7 @@ public:
 
       (const String &action,
        const Properties &params,
-       const Properties &globdat);
+       const Properties &globdat) override;
 
   /// @brief Factory method for creating new HingeModel instances
   /// @param name Model name
