### 2026-05-24 dfcebc3
fixed unused-parameters warning (cf OBS-09)
