+++ src/misc/CSVDataPrinter.cpp
@@ -61,7 +61,7 @@ void CSVDataPrinter::printTable
   {
     mode_ = TABLES;
     writePrefix_(out, globdat);
-    startTables_(out, label, table, globdat);
+    startTables_(out, table, globdat);
   }
   JEM_PRECHECK2(mode_ == TABLES, "Cannot Output Tables and Vectors in the same csv, try using seperate modules!");
 
@@ -96,7 +96,7 @@ void CSVDataPrinter::printVector
   {
     mode_ = VECTORS;
     writePrefix_(out, globdat);
-    startVectors_(out, label, vec, dofs, globdat);
+    startVectors_(out, dofs, globdat);
   }
   JEM_PRECHECK2(mode_ == VECTORS, "Cannot Output Tables and Vectors in the same csv, try using seperate modules!");
 
@@ -122,7 +122,7 @@ void CSVDataPrinter::printSparseVector
   {
     mode_ = SPARSEVECTORS;
     writePrefix_(out, globdat);
-    startVectors_(out, label, vec, dofs, globdat, idofs);
+    startVectors_(out, dofs, globdat, idofs);
   }
   JEM_PRECHECK2(mode_ == SPARSEVECTORS, "Cannot Output Tables and Vectors in the same csv, try using seperate modules!");
 
@@ -137,11 +137,12 @@ void CSVDataPrinter::printSparseVector
 void CSVDataPrinter::startTables_
 
     (Output &out,
-     const String &label,
      const Table &table,
      const Properties &globdat)
 
 {
+  (void)globdat; // unused
+
   double val;
   IdxVector rows(table.rowCount());
   table.getRowItems()->getItemMap()->getItemIDs(rows, IdxVector(jem::iarray(table.rowCount())));
@@ -161,13 +162,13 @@ void CSVDataPrinter::startTables_
 void CSVDataPrinter::startVectors_
 
     (Output &out,
-     const String &label,
-     const Vector &vec,
      const DofSpace &dofs,
      const Properties &globdat,
      const IdxVector &idofs)
 
 {
+  (void)globdat; // unused
+
   IdxVector reportDofs = idofs.size() ? idofs : IdxVector(jem::iarray(dofs.dofCount()));
 
   for (idx_t dof : reportDofs)
