+++ src/modules/MilneDeviceModule.cpp
@@ -147,6 +147,10 @@ Ref<Module> MilneDeviceModule::makeNew(const String &name, const Properties &con
                                        const Properties &props, const Properties &globdat)
 
 {
+  (void)conf;    // unused
+  (void)props;   // unused
+  (void)globdat; // unused
+
   return newInstance<Self>(name);
 }
 
