+++ src/modules/GMSHInputModule.cpp
@@ -545,6 +545,10 @@ Ref<Module> GMSHInputModule::makeNew
      const Properties &globdat)
 
 {
+  (void)conf;    // unused
+  (void)props;   // unused
+  (void)globdat; // unused
+
   return jem::newInstance<Self>(name);
 }
 
