+++ src/modules/ExplicitModule.cpp
@@ -303,7 +303,7 @@ bool ExplicitModule::commit(const Properties &globdat)
   {
     params.find(accept, ActionParams::ACCEPT);
   }
-  accept = accept && (error <= prec_ || dtime_ == minDtime_);
+  accept = accept && (error <= prec_ || dtime_ <= minDtime_);
 
   if (accept)
   {
@@ -323,9 +323,9 @@ bool ExplicitModule::commit(const Properties &globdat)
   }
 
   jem::System::info(myName_) << " ...Adapting time step size to " << dtime_ << "\n";
-  if (dtime_ == maxDtime_ && dtime_ > minDtime_)
+  if (dtime_ >= maxDtime_ && dtime_ > minDtime_)
     jem::System::info(myName_) << " !!! Largest allowed time step !!!\n";
-  if (dtime_ == minDtime_ && dtime_ < maxDtime_)
+  if (dtime_ <= minDtime_ && dtime_ < maxDtime_)
     jem::System::info(myName_) << " !!! Smallest allowed time step !!!\n";
   Globdat::getVariables(globdat).set(jive::implict::PropNames::DELTA_TIME,
                                      dtime_);
