### 2026-05-24 1d969f9
makefile with optional stricter warning hygiene
