### 2026-05-24 aa50dab
replace Float::EPSILON with jem::Float::EPSILON to fix a lot of errors
