+++ src/materials/ElastoPlasticRodMaterial.cpp
@@ -10,6 +10,7 @@
 #include "utils/testing.h"
 
 #include <jem/base/Exception.h>
+#include <jem/base/Float.h>
 #include <jem/base/ClassTemplate.h>
 #include <jem/util/PropertyException.h>
 #include <jem/util/StringUtils.h>
@@ -216,7 +217,7 @@ void ElastoPlasticRodMaterial::getStress(const Vector &stress, const Vector &str
     Super::getStress(stress, Vector(strain - plastStrain), ielem, ip, false);
     getHardVals(hardStress, hardParams);
 
-    if (!inelastic || ((jem::numeric::abs(jem::edgeFact_ - 1.0) > Float::EPSILON) && (ielem < edgeElems_ || ielem > nElem_ - edgeElems_ - 1)))
+    if (!inelastic || ((jem::numeric::abs(edgeFact_ - 1.0) > jem::Float::EPSILON) && (ielem < edgeElems_ || ielem > nElem_ - edgeElems_ - 1)))
     {
       if (verbosity_ > 1)
         jem::System::debug(myName_) << "        elastic calculation\n";
@@ -247,7 +248,7 @@ void ElastoPlasticRodMaterial::getStress(const Vector &stress, const Vector &str
     {
       yieldGrad = jive_helpers::funcGrad(yieldCond_, args);
       for (idx_t i = 0; i < dofCount_; i++)
-        if (jem::numeric::abs(args[i]) < Float::EPSILON)
+        if (jem::numeric::abs(args[i]) < jem::Float::EPSILON)
           yieldGrad[i] = 0.;
     }
 
@@ -291,7 +292,7 @@ void ElastoPlasticRodMaterial::applyDeform()
       energyPot_(ip, ielem) = 0.5 * dotProduct(currElastStrain, currStress);
 
       WARN_ASSERT2(currDeltaFlow_(ip, ielem) >= 0., "Negative plastic multiplier");
-      if (jem::numeric::abs(currDeltaFlow_(ip, ielem)) > Float::EPSILON)
+      if (jem::numeric::abs(currDeltaFlow_(ip, ielem)) > jem::Float::EPSILON)
       {
         energyDiss_(ip, ielem) += dotProduct((oldStress + currStress) / 2., deltaPlastStrain);
       }
