### 2026-05-25 093036b
Update issue templates
