+++ .github/ISSUE_TEMPLATE/numerical-bug.md
@@ -1,0 +1,20 @@
+---
+name: Numerical Bug
+about: Report any numerical issues you encounter
+title: "[BUG]"
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Please provide a minimal `.pro` file together with all other inputs needed to reproduce the bug. Also what version of dynLattice are you using? 
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Additional context**
+Add any other context about the problem here.
