+++ .github/pull_request_template.md
@@ -1,0 +1,18 @@
+## Summary
+
+Describe what was changed
+
+## Validation
+
+- [ ] relevant benchmarks were added (if applicable) 
+- [ ] all tests/benchmarks run succesfully 
+- [ ] documentation is added to all files
+- [ ] the code compiles without project related warnings
+
+## Scientific context
+
+List any relevant benchmark, analytical result, or literature reference if applicable.
+
+## Related issue
+
+Closes #
