### 2026-05-28 cbbc648
added option for stricter warning hygiene as well as an CI job to check for it. Introduced one warning for testing purposes
