+++ .github/workflows/hygienic-build.yml
@@ -1,0 +1,40 @@
+name: Hygienic Build
+on:
+  push: 
+    paths:
+      - src/**
+      - .github/workflows/hygienic-build.yml
+      - Makefile
+
+jobs:
+  hygienic-build:
+    runs-on: ubuntu-latest
+    name: hygiene test build
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v5
+      - name: Apptainer
+        uses: eWaterCycle/setup-apptainer@v2
+        with:
+          apptainer-version: 1.4.2
+      - name: Download jive.sif
+        run: wget -O jive.sif "https://github.com/hortulanusT/dynLattice/raw/${GITHUB_SHA}/jive.sif" && echo "Downloaded jive.sif for commit ${GITHUB_SHA}"
+      - name: Build hygienic
+        run: |
+          apptainer exec jive.sif make HYGIENE=1 > hygienic_build.log 2>&1
+          grep "warning:" hygienic_build.log | grep "^src/" | sort -u > hygienic_warnings.log || true
+
+          if [[ -s hygienic_warnings.log ]]; then
+            echo "::error::Hygienic build produced project warnings"
+            echo "Warnings found:"
+            sed -n 's/.*\[-W\([^]]*\)\].*/\1/p' hygienic_warnings.log | sort | uniq -c | sort -nr
+            exit 1
+          fi
+      - name: Upload hygienic warnings
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: hygienic-warnings
+          path: |
+            hygienic_build.log
+            hygienic_warnings.log
