+++ Makefile
@@ -23,26 +23,27 @@ MY_INCDIRS 	:= $(SRCDIR)
 #######################################################################
 ##   Defining C++ Standard & stricter warnings                       ##
 #######################################################################
-# Add warnings one by one by uncommenting lines.
+# Default build: regular warnings only.
+# Hygienic build: enable stricter warning hygiene with:
+#   make HYGIENE=1
+HYGIENE ?= 0
+
 MY_WARN_FLAGS :=
-
-### UNCOMMENT THE LINES BELOW FOR STRICTER WARNING HYGIENE
-# MY_WARN_FLAGS += '-Wextra'
-# MY_WARN_FLAGS += '-Wpedantic'
-# MY_WARN_FLAGS += '-Wshadow'
-# MY_WARN_FLAGS += '-Wsuggest-override'
-# MY_WARN_FLAGS += '-Wold-style-cast'
-# MY_WARN_FLAGS += '-Woverloaded-virtual'
-# MY_WARN_FLAGS += '-Wdouble-promotion'
-
-# MY_WARN_FLAGS += '-Wundef'
-# MY_WARN_FLAGS += '-Wnon-virtual-dtor'
-# MY_WARN_FLAGS += '-Wundef'
-# MY_WARN_FLAGS += '-Winit-self'
-
-# MY_WARN_FLAGS += '-Wconversion'
-# MY_WARN_FLAGS += '-Wsign-conversion'
-# MY_WARN_FLAGS += '-Wfloat-equal'
+ifeq ($(HYGIENE),1)
+MY_WARN_FLAGS += '-Wextra'
+MY_WARN_FLAGS += '-Wpedantic'
+MY_WARN_FLAGS += '-Wshadow'
+MY_WARN_FLAGS += '-Wsuggest-override'
+MY_WARN_FLAGS += '-Wold-style-cast'
+MY_WARN_FLAGS += '-Woverloaded-virtual'
+MY_WARN_FLAGS += '-Wdouble-promotion'
+MY_WARN_FLAGS += '-Wundef'
+MY_WARN_FLAGS += '-Wnon-virtual-dtor'
+MY_WARN_FLAGS += '-Winit-self'
+MY_WARN_FLAGS += '-Wconversion'
+MY_WARN_FLAGS += '-Wsign-conversion'
+MY_WARN_FLAGS += '-Wfloat-equal'
+endif
 
 MY_CXX_STD_FLAGS := '-std=c++17' $(MY_WARN_FLAGS)
 #######################################################################
@@ -59,7 +60,7 @@ MKDIR_P	= mkdir -p --
 include ${shell find $(SRCDIR) -name *.mk}
 
 #######################################################################
-##   add opemMPI libraries                                           ##
+##   add openMPI libraries                                           ##
 #######################################################################
 export LD_RUN_PATH := $(MPIDIR)/lib:$(LD_RUN_PATH)
 
