+++ .github/workflows/build-docs.yml
@@ -13,7 +13,7 @@ jobs:
     name: Doxygen Documentation
     steps:
       - name: Checkout
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
       - name: Apptainer
         uses: eWaterCycle/setup-apptainer@v2
         with:
@@ -23,7 +23,7 @@ jobs:
       - name: Build documentation
         run: apptainer exec jive.sif make docs
       - name: Upload documentation
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: documentation
           path: doc/html
@@ -35,12 +35,12 @@ jobs:
     if: github.ref == 'refs/heads/main'
     steps:
       - name: Download documentation
-        uses: actions/download-artifact@v5
+        uses: actions/download-artifact@v8
         with:
           name: documentation
           path: doc/html
       - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v3
+        uses: peaceiris/actions-gh-pages@v4
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: doc/html
