+++ .github/workflows/hygienic-build.yml
@@ -12,7 +12,7 @@ jobs:
     name: hygiene test build
     steps:
       - name: Checkout
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
       - name: Apptainer
         uses: eWaterCycle/setup-apptainer@v2
         with:
@@ -32,7 +32,7 @@ jobs:
           fi
       - name: Upload hygienic warnings
         if: always()
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: hygienic-warnings
           path: |
