### 2026-05-28 982e8c0
added some documentation for the warning checks
