### 2026-05-31 86ad5f4
Added Release CI & Updated Zenodo DOI (cf OBS-07)
