+++ .github/workflows/release-on-tag.yml
@@ -1,0 +1,38 @@
+name: Release On Tag
+
+on:
+  push:
+    tags:
+      - "v*"
+
+permissions:
+  contents: write
+
+jobs:
+  release:
+    runs-on: ubuntu-latest
+    name: Create release assets
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+        with:
+          fetch-depth: 0
+          lfs: true
+      - name: Create source archive from tag
+        run: |
+          TAG="${GITHUB_REF_NAME}"
+          git archive --format=zip --output "dynLattice-${TAG}.zip" "${TAG}" -- . ':!jive.sif' ':!paper'  ':!.github'
+      - name: Verify release assets exist
+        run: |
+          test -f "dynLattice-${GITHUB_REF_NAME}.zip"
+          test -f jive.sif
+      - name: Publish GitHub release
+        uses: softprops/action-gh-release@v2
+        with:
+          tag_name: ${{ github.ref_name }}
+          name: ${{ github.ref_name }}
+          files: |
+            dynLattice-${{ github.ref_name }}.zip
+            jive.sif
+          fail_on_unmatched_files: true
+          overwrite_files: true
