### 2026-05-31 23a72cb
OBS-03/07/10: upgrade beam benchmarks to numerical regression tests
