+++ tests/beam/test5.py
@@ -1,9 +1,16 @@
 #!/usr/bin/python3
 
 # TEST 5 3D
+import sys
 import numpy as np
+from pathlib import Path
 from termcolor import colored
 from matplotlib import pyplot as plt
+
+sys.path.insert(0, str(Path(__file__).parent.parent))
+from metrics import interp_on_reference, relative_L2
+
+TOL = 0.05
 
 test_passed = False
 
@@ -31,16 +38,26 @@ try:
   plt.xlabel("load (N)")
   plt.ylabel("displacement (m)")
   plt.xlim(left=0, right=3000)
+
+  # Separate check for each tip displacement component 
+  load = sim_resp[:, 2]
+  u1 = interp_on_reference(load, sim_disp[:, 0], ref_data_u1[:, 0])
+  u2 = interp_on_reference(load, sim_disp[:, 1], ref_data_u2[:, 0])
+  u3 = interp_on_reference(load, sim_disp[:, 2], ref_data_u3[:, 0])
+  err_u1 = relative_L2(u1, ref_data_u1[:, 1])
+  err_u2 = relative_L2(u2, ref_data_u2[:, 1])
+  err_u3 = relative_L2(u3, ref_data_u3[:, 1])
+  test_passed = all(e <= TOL for e in [err_u1, err_u2, err_u3])
+
 except Exception as e:
   print(e)
-else:
-  test_passed = True
 
 if test_passed:
-  print(colored("STATIC TEST 5 RUN THROUGH", "green"))
+  print(colored("STATIC TEST 5 PASSED", "green"))
 
   plt.tight_layout()
   plt.savefig("tests/beam/test5/result.pdf")
   plt.savefig("tests/beam5_result.png")
 else:
   print(colored("STATIC TEST 5 FAILED", "red", attrs=["bold"]))
+  sys.exit(1)
