### 2026-05-31 26a5776
OBS-03/08/10: upgrade plastic benchmarks to numerical regression tests
no test for plastic/test3.py, as here only the real FE² data is recorded, this is more a qualitiative test to show, that the results work in general for the documentation
