### 2026-05-31 11053ad
Added Test Cases to CI and cleaned up the CI workflows
