+++ .github/workflows/build-docs.yml
@@ -1,46 +1,0 @@
-name: Build Docs
-on:
-  push: 
-    paths:
-      - src/**
-      - doc/**
-      - tests/*.md
-      - .github/workflows/build-docs.yml
-
-jobs:
-  documentation:
-    runs-on: ubuntu-latest
-    name: Doxygen Documentation
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v6
-      - name: Apptainer
-        uses: eWaterCycle/setup-apptainer@v2
-        with:
-          apptainer-version: 1.4.2
-      - name: Download jive.sif
-        run: wget -O jive.sif "https://github.com/hortulanusT/dynLattice/raw/${GITHUB_SHA}/jive.sif" && echo "Downloaded jive.sif for commit ${GITHUB_SHA}"
-      - name: Build documentation
-        run: apptainer exec jive.sif make docs
-      - name: Upload documentation
-        uses: actions/upload-artifact@v7
-        with:
-          name: documentation
-          path: doc/html
-
-  gh-pages:
-    runs-on: ubuntu-latest
-    name: Deploy to GitHub Pages
-    needs: documentation
-    if: github.ref == 'refs/heads/main'
-    steps:
-      - name: Download documentation
-        uses: actions/download-artifact@v8
-        with:
-          name: documentation
-          path: doc/html
-      - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v4
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: doc/html
