+++ .github/workflows/deploy-docs.yml
@@ -1,0 +1,37 @@
+name: Deploy Docs
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  build-and-deploy:
+    runs-on: ubuntu-latest
+    name: Build and deploy documentation
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+      - name: Apptainer
+        uses: eWaterCycle/setup-apptainer@v2
+        with:
+          apptainer-version: 1.4.2
+      - name: Compute jive.sif hash
+        id: jive-hash
+        run: echo "hash=$(git rev-parse HEAD:jive.sif)" >> $GITHUB_OUTPUT
+      - name: Restore jive.sif
+        id: cache-jive
+        uses: actions/cache@v4
+        with:
+          path: jive.sif
+          key: jive-sif-${{ steps.jive-hash.outputs.hash }}
+      - name: Download jive.sif
+        if: steps.cache-jive.outputs.cache-hit != 'true'
+        run: wget -O jive.sif "https://github.com/hortulanusT/dynLattice/raw/${{ github.sha }}/jive.sif"
+      - name: Build documentation
+        run: apptainer exec jive.sif make docs
+      - name: Deploy to GitHub Pages
+        uses: peaceiris/actions-gh-pages@v4
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: doc/html
