+++ .github/workflows/release-on-tag.yml
@@ -15,9 +15,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v6
-        with:
-          fetch-depth: 0
-          lfs: true
+      - name: Download jive.sif
+        run: wget -O jive.sif "https://github.com/hortulanusT/dynLattice/raw/${GITHUB_SHA}/jive.sif" && echo "Downloaded jive.sif for commit ${GITHUB_SHA}"
       - name: Create source archive from tag
         run: |
           TAG="${GITHUB_REF_NAME}"
