+++ search/all_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['4_0',['Test 4',['../beambenchmarks.html#autotoc_md52',1,'']]]
+  ['4_0',['Test 4',['../beambenchmarks.html#autotoc_md55',1,'']]]
 ];

[…]
