### 2026-08-03 hortulanusT/dynLattice#15 @mohd-afeef-badri
Installation help
Dear  Developers,

I am here since I am reviewing your JOSS paper  https://github.com/openjournals/joss-reviews/issues/10410#

Firstly, sorry for being late.
Secondly I am new to Apptainer so you would need to help me.

I tried to install the Apptainer image (jive.sif) on my UB24.04 but I get error

```
$ jive make
make: *** No targets specified and no makefile found.  Stop.
```

I verified that apptainer is installed

```
$ apptainer version
1.5.0
```

even I was able to inspect your `.sif` file

```
$ apptainer inspect jive.sif
creator: t.gartner@tudelft.nl
org.label-schema.build-arch: amd64
org.label-schema.build-date: Friday_8_August_2025_10:37:39_CEST
org.label-schema.schema-version: 1.0
org.label-schema.usage: /.singularity.d/runscript.help
org.label-schema.usage.apptainer.runscript.help: /.singularity.d/runscript.help
org.label-schema.usage.apptainer.version: 1.4.1
org.label-schema.usage.singularity.deffile.bootstrap: docker
org.label-schema.usage.singularity.deffile.from: ubuntu
org.opencontainers.image.ref.name: ubuntu
org.opencontainers.image.version: 24.04
```

can you please help.
