### 2026-08-03 hortulanusT/dynLattice#16 @mohd-afeef-badri
Tutorials
Dear Developers,

I am writing as part of the JOSS review process  (see: https://github.com/openjournals/joss-reviews/issues/10410#).

After exploring the resources provided with **dynLattice**, my understanding is that the only end-user tutorial currently available is the Usage page here: https://gaertner.cv/dynLattice/usage.html .

Since the software is intended to produce numerical simulations, I believe the tutorial could be significantly improved by providing more context about the benchmark example, including the underlying physics and the purpose of the test. Adding figures that illustrate the geometry and examples of the ParaView post-processing results would also help users understand what to expect from the simulation outputs.

In addition, I suggest including two or three additional example cases to better demonstrate the capabilities of the software. A broader set of examples would make it easier for new users to understand potential applications and would help motivate the open-source community to adopt, contribute to, and further develop the software.

As part of the JOSS review, I am opening this issue to document these suggestions for improving the user documentation.
