### 2026-08-24 3230e89
addressing point 1 of https://github.com/hortulanusT/dynLattice/issues/16
better documentation of the used example case (transient 1)
