### 2026-08-24 hortulanusT/dynLattice#15 @hortulanusT
Hi @mohd-afeef-badri ,

thank you for your patience, being back from vacation I now tried to reproduce your error on multiple machines, but was not able to encounter the behavior, you describe. Just to be sure, you downloaded both the container *and* the accompanying source code? Are you running this command from the folder you extracted the latest release/cloned the repository into?
