### 2026-08-25 c3cafb5
Update installation instructions
