### 2026-08-27 hortulanusT/dynLattice#15 @mohd-afeef-badri
@hortulanusT

Thank you for your response. Indeed it now works I was only downloading the `.sif` before.  When I untared your release repository, I was able to run the `jive make`. and the tests:

```
STATIC TEST 1 PASSED
CONTACT TEST 1 RUN THROUGH
PLASTIC TEST 2b RUN THROUGH
PLASTIC TEST 2a RUN THROUGH
STATIC TEST 5 RUN THROUGH
PLASTIC TEST 1 RUN THROUGH
PLASTIC TEST 3 RUN THROUGH
STATIC TEST 2 RUN THROUGH
TRANSIENT TEST 2 RUN THROUGH
STATIC TEST 4 RUN THROUGH
TRANSIENT TEST 3 RUN THROUGH
```

Perhaps it will be great for people not aware of apptainer use to explicity mention this to "download the conatiner (`.sif`) and source code".
