### 2026-08-27 db68dcb
Add schematic visualization and explanation for a second example: a re-entrant honeycomb impact
