+++ doc/usage_details1.md
@@ -1,0 +1,135 @@
+\page prop1 Properties for Example 1
+
+| Command | Explanation |
+| ------- | ----------- |
+| **Program Control** | |
+| `control.runWhile = "t <= 30";` | lets the control module (cf. main.cpp) terminate the simulation once the variable `t` (defaults to the current simulation time in seconds) has become greater than `30`. |
+| **Solver Configuration** | |
+| `Solver.modules = [ "integrator" ];` | defines the modules, that make up the solver. In our case it is only one, called `integrator`. |
+| `Solver.integrator.type = "MilneDevice";` | defines the type of the `integrator` module, in our case it's a MilneDeviceModule (this type needs to agree with the name it is declared under in the [ModuleFactory](https://jive-manual.dynaflow.com/classjive_1_1app_1_1ModuleFactory.html)) |
+| `Solver.integrator.deltaTime = 5e-5;` | set the starting time step size for the `integrator` to \f$5 \times 10^{-5}\f$ seconds |
+| **Material Properties** | |
+| `params.rod_details.material.type = "ElasticRod";` | define the type of the material to be ElasticRodMaterial |
+| `params.rod_details.material.cross_section = "square";` | define the rod to be of square cross-section |
+| `params.rod_details.material.side_length = "sqrt(12/2e3)";` | define the sides to be \f$\sqrt{12/2e3}\approx0.0775\f$ (meter) long |
+| `params.rod_details.material.young = "5.6e10/12";` | define the rods Young's modulus to be \f$5.6e10/12\approx4.67e9\f$ (Pascal) |
+| `params.rod_details.material.shear_modulus = 2e9;` | define the rods shear modulus to be \f$2e9\f$ (Pascal) |
+| `params.rod_details.material.density = 200.;` | define the rod to have a density of \f$200\f$ (kilogram/meter³) |
+| **File Includes** | |
+| `include "input.pro";` | include the \ref input file defining input modules and geometry reading |
+| `include "model.pro";` | include the \ref model file defining the physical model |
+| `include "output.pro";` | include the \ref output file defining output modules and data collection |
+| **Additional Settings** | |
+| `Input.input.order = 2;` | set the interpolation order for the input elements to 2 (quadratic elements) |
+| **Force Model Override** | |
+| `model.model.force.type = "None";` | override the force model from model.pro to have no applied forces |
+| **Displacement Boundary Conditions** | |
+| `model.model.disp.type = "LoadScale";` | set the displacement boundary condition to use load scaling ([LoadScaleModel](https://jive-manual.dynaflow.com/classjive_1_1model_1_1LoadScaleModel.html)) (in dynamic simulations displacement prescribtions are prescribing accelerations) |
+| `model.model.disp.scaleFunc = "if (t<15, 6/15 * (1 - cos(2*PI/15 * t)), 0)";` | define a time-dependent scaling function: a single sinusoidal half-wave during the first 15s |
+| `model.model.disp.model.type = "Dirichlet";` | specify that the displacement model uses Dirichlet boundary conditions (DirichletModel) (they will apply to the acceleration in the dynamic solves) |
+| `model.model.disp.model.nodeGroups = [ "fixed" ];` | apply the displacement (acceleration) boundary condition to the "fixed" node group |
+| `model.model.disp.model.factors = [ 1. ];` | set the scaling factor for the displacement (acceleration) to 1.0 |
+| `model.model.disp.model.dofs = [ "rz" ];` | apply the displacement (acceleration) boundary condition to the rz (rotation about z-axis) degree of freedom |
+| **CSV Output Configuration** | |
[…]

[…]
