### 2026-08-27 a685c65
Update installation instructions for clarity and consistency in line with #17
