### 2026-08-27 ed47b18
Update documentation and model parameters for Test 2 simulation
