+++ doc/usage.md
@@ -31,9 +31,9 @@ Test 2 is adapted from the impact experiments in [Gärtner et al. (2025)](https:
 
 The top edge represents a rigid impactor: it carries an attached mass and is given an initial downward velocity of \f$v_0=70\f$ m/s, so it starts flying freely into the lattice below it. The bottom edge rests on a compliant support, modeled as a slender elastic spring rod representing the finite stiffness of the support/load-cell plate used in the experiments. Both edges are otherwise guided to move only vertically. Since the lattice walls come into contact with each other as the re-entrant cells collapse, both rod-rod and rod-joint self-contact are enabled. 
 
-The simulation runs until the impactor has moved down by half the lattice's height and its velocity becomes non-negative again, i.e. until the lattice has been crushed by about 50% and starts to rebound. Physically, a compressive wave runs down through the tiled unit cells, with plastic hinges forming at the rod junctions (visible as the darker segments in the animation below) and the re-entrant cells drawing inward rather than bulging outward, which is the defining (negative Poisson's ratio) characteristic of auxetic geometries. Unlike Test 1, this case does not have literature reference data to compare against directly here; it primarily illustrates a more complex model setup, see \ref transientbenchmarks and \ref contactbenchmarks for validated benchmarks of the individual ingredients (contact, plasticity) used here.
+The simulation runs until the impactor has moved down by half the lattice's height and its velocity becomes non-negative again, i.e. until the lattice has been crushed by about 50% and starts to rebound. Physically, a compressive wave runs down through the tiled unit cells, with plastic hinges forming at the rod junctions (visible as the brighter segments around the (elastic) joints) and the re-entrant cells drawing inward rather than bulging outward, which is the defining characteristic of auxetic geometries. Unlike Test 1, there is no direct literature data to compare to, but the corresponding publication [Gärtner et al. (2025)](https://doi.org/10.1016/j.ijimpeng.2025.105402) compares a similar (but larger) setup against Experiments and commercial FE-Engines; it primarily illustrates a more complex model setup, see \ref transientbenchmarks and \ref contactbenchmarks for validated benchmarks of the individual ingredients (contact, plasticity) used here.
 
-![Test 2 Animation](test2_animate.gif)
+![Test 2 Animation (1 Frame ~ 1% Strain)](test2_animate.gif)
 
 ## Setup
 The workflow is the same as for \ref transient1 "Example 1": compile with `jive make`, then supply a geometry file (here `tests/docs/re-entrant.geo`, parameterized so the same file generates the honeycomb for any cell angle/aspect ratio/repetition count) and a property file (`tests/docs/test2.pro`, including `input.pro`, `model.pro` and `output.pro` from the same folder). A detailed, line-by-line overview of the properties can be found in \ref prop2.
