### 2026-08-27 912383e
Refactor example 2 parameters and remove obsolete graphics
