### 2026-08-27 #10410 @hortulanusT
@mbarzegary Please excuse my long silence. We had the chance to address the two issues raised, for https://github.com/hortulanusT/dynLattice/issues/15 @fmeer cleared up the installation instructions and made the requirements around the apptainer installation more clear. For https://github.com/hortulanusT/dynLattice/issues/16 we added for two examples (one more academic and one from a real application) into the documentation, with more illustration and more background. I have to admit this last change is still not fully approved both internally and from @mohd-afeef-badri, and thus also the PR https://github.com/hortulanusT/dynLattice/pull/18 is not merged yet.

One other question, that came up, when I saw, that there are quite some commits since I put the repo forward to the Review now and no new release, will there be a "accepted release", that includes all the changes resulting from the Review? Is there any formal process for this, or do I simply bump the version number and inform @editorialbot ?

Thank you again for your help! If there is anything else, you need from my side, please let me know!
