### 2026-08-27 hortulanusT/dynLattice#15 @hortulanusT
Hi @mohd-afeef-badri ,

we updated the installation instructions in both the doc/install.md and the main README file, thank you for pointing out this unclear language from our side!

I would close this issue after the PR unless you object.
