### 2026-08-27 hortulanusT/dynLattice#16 @hortulanusT
Hi @mohd-afeef-badri

thank you for raising this issue. We have adapted the documentation to include two examples, (one from the test cases as previously and one from the application paper the code was used in). We have, for both cases included a graphic for the Setup, a GIF for the Results and a separate page, where we explain the settings in more detail.

I am not fully sure how you can see these changes best, as I think they will not be visible in the public Github Pages until we merge #18, would you think it is easier for you, if I just upload the images/GIFs in this issue and let you read the updated doc/usage*.md or would you prefer to clone the repo and build a local version of the documentation?

Thank you again for your review!
