### 2026-08-28 hortulanusT/dynLattice#15 @mohd-afeef-badri
@hortulanusT
Thanks for adding the installation instructions. Feel free to close this issue 👍🏻
