### 2026-08-28 hortulanusT/dynLattice#16 @mohd-afeef-badri
@hortulanusT @fmeer

Thanks fir adapting the documentation. I am sure this will be of use to someone who would like to understand more on the workings of your code.

Concerning your question, about reviewing the work done on the tutorials, I can try to clone the repo and build the local version of documentation.

Do you have instructions on how to build the documentation ?
