### 2026-08-28 hortulanusT/dynLattice#16 @hortulanusT
@mohd-afeef-badri

yes, the documentation is build by the make target `docs`, so after cloning the repo and switching to my development branch, you can simply execute `jive make docs` and it should fully work. You can find then the documentation starting page under `doc/html/index.html`
