### 2026-08-29 hortulanusT/dynLattice#16 @mohd-afeef-badri
@hortulanusT and @fmeer

Thanks for making the efforts for the two examples you have provided in the Usage page. I was able to go through them and follow the two examples which I feel are nicely explained. I am sure this would be helpful for the project in future.

Please feel free to close this issue after you merge the branch into the main.
