+++ .github/workflows/ci.yml
@@ -17,7 +17,7 @@ jobs:
       - uses: actions/checkout@v6
       - id: jive-hash
         run: echo "hash=$(git rev-parse HEAD:jive.sif)" >> $GITHUB_OUTPUT
-      - uses: actions/cache@v4
+      - uses: actions/cache@v6
         id: cache-jive
         with:
           path: jive.sif
@@ -38,7 +38,7 @@ jobs:
           apptainer-version: 1.4.2
       - id: jive-hash
         run: echo "hash=$(git rev-parse HEAD:jive.sif)" >> $GITHUB_OUTPUT
-      - uses: actions/cache@v4
+      - uses: actions/cache@v6
         with:
           path: jive.sif
           key: jive-sif-${{ steps.jive-hash.outputs.hash }}
@@ -67,7 +67,7 @@ jobs:
           apptainer-version: 1.4.2
       - id: jive-hash
         run: echo "hash=$(git rev-parse HEAD:jive.sif)" >> $GITHUB_OUTPUT
-      - uses: actions/cache@v4
+      - uses: actions/cache@v6
         with:
           path: jive.sif
           key: jive-sif-${{ steps.jive-hash.outputs.hash }}
