+++ search/pages_1.js
@@ -1,7 +1,4 @@
 var searchData=
 [
-  ['code_20of_20conduct_0',['Code of Conduct',['../md_CODE__OF__CONDUCT.html',1,'']]],
-  ['conduct_1',['Code of Conduct',['../md_CODE__OF__CONDUCT.html',1,'']]],
-  ['contact_20benchmarks_2',['Contact Benchmarks',['../contactbenchmarks.html',1,'benchmark']]],
-  ['contributing_20to_20dynlattice_3',['Contributing to dynLattice',['../md_CONTRIBUTING.html',1,'']]]
+  ['2_0',['Properties for Example 2',['../prop2.html',1,'']]]
 ];
