+++ search/pages_5.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['plastic_20benchmarks_0',['Plastic Benchmarks',['../plasticbenchmarks.html',1,'benchmark']]]
+  ['example_201_0',['Properties for Example 1',['../prop1.html',1,'']]],
+  ['example_202_1',['Properties for Example 2',['../prop2.html',1,'']]]
 ];
