+++ search/pages_0.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['beam_20benchmarks_0',['Beam Benchmarks',['../beambenchmarks.html',1,'benchmark']]],
-  ['benchmarks_1',['benchmarks',['../beambenchmarks.html',1,'Beam Benchmarks'],['../benchmark.html',1,'Benchmarks'],['../contactbenchmarks.html',1,'Contact Benchmarks'],['../plasticbenchmarks.html',1,'Plastic Benchmarks'],['../transientbenchmarks.html',1,'Transient Benchmarks']]]
+  ['1_0',['Properties for Example 1',['../prop1.html',1,'']]]
 ];
