+++ search/all_2.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['3_0',['3',['../plasticbenchmarks.html#autotoc_md65',1,'Test 3'],['../transientbenchmarks.html#autotoc_md68',1,'Test 3']]],
-  ['3_20using_20the_20code_1',['3. Using the code',['../install.html#autotoc_md24',1,'']]]
+  ['3_0',['3',['../plasticbenchmarks.html#autotoc_md70',1,'Test 3'],['../transientbenchmarks.html#autotoc_md73',1,'Test 3']]],
+  ['3_20extend_20the_20code_20or_20use_20it_20for_20new_20cases_1',['3. Extend the code or use it for new cases',['../install.html#autotoc_md23',1,'']]]
 ];
