+++ search/all_13.js
@@ -6,22 +6,23 @@ var searchData=
   ['name_3',['name',['../structParaViewModule_1_1ElementInfo.html#a7ff5625e713c30a5760d8e29d25d90fe',1,'ParaViewModule::ElementInfo']]],
   ['name_5fprefix_4',['NAME_PREFIX',['../classLatticeModel.html#a5dfed6c245fd8df63d0bc181a3e94d84',1,'LatticeModel']]],
   ['nametovtknum_5',['nameToVTKNum',['../classParaViewModule.html#a908dc275cea81a14378eb5f730db2c94',1,'ParaViewModule']]],
-  ['need_6',['Statement of need',['..//home/runner/work/dynLattice/dynLattice/paper/paper.md#autotoc_md32',1,'']]],
+  ['need_6',['Statement of need',['..//home/runner/work/dynLattice/dynLattice/paper/paper.md#autotoc_md37',1,'']]],
   ['nelem_5f_7',['nElem_',['../classElasticRodMaterial.html#ab853faec779c62ca78f496c5f1833435',1,'ElasticRodMaterial']]],
   ['neumannmodel_8',['neumannmodel',['../classNeumannModel.html',1,'NeumannModel'],['../classNeumannModel.html#ab9a9a19d2f8c0f14489c111656cb05f0',1,'NeumannModel::NeumannModel()']]],
   ['neumannmodel_2ecpp_9',['NeumannModel.cpp',['../NeumannModel_8cpp.html',1,'']]],
   ['neumannmodel_2eh_10',['NeumannModel.h',['../NeumannModel_8h.html',1,'']]],
-  ['new_20components_11',['Adding New Components',['../md_CONTRIBUTING.html#autotoc_md7',1,'']]],
-  ['newinstance_12',['newinstance',['../classMaterialFactory.html#a29f0b52914e228f01fc150cd5ce5f4eb',1,'MaterialFactory::newInstance(const String &amp;name, const Properties &amp;conf, const Properties &amp;props, const Properties &amp;globdat)'],['../classMaterialFactory.html#a3dc34255c886e833e0147c5e85d08431',1,'MaterialFactory::newInstance(const String &amp;type, const String &amp;name, const Properties &amp;conf, const Properties &amp;props, const Properties &amp;globdat)']]],
-  ['ngroups_5fprop_13',['NGROUPS_PROP',['../classPBCGroupInputModule.html#ace97da1c76f02a9de922eaef9bdeaa27',1,'PBCGroupInputModule']]],
-  ['node_5fgroups_14',['NODE_GROUPS',['../classGroupInputModule.html#a053e0d3126619830fd06ad821f8c9c4b',1,'GroupInputModule']]],
-  ['node_5fgroups_5fprop_15',['NODE_GROUPS_PROP',['../classMassModel.html#a0cca2745bf365a54d08acf60eff7309d',1,'MassModel']]],
-  ['node_5fprop_16',['NODE_PROP',['../classForceOutputModule.html#a3cdecec8917e3a7866969086179dcf4e',1,'ForceOutputModule']]],
-  ['nodecount_17',['nodeCount',['../classLine3D.html#a65457d2ea9cd00c9514b9637463fff69',1,'Line3D']]],
-  ['nodedata_18',['nodeData',['../structParaViewModule_1_1ElementInfo.html#a2acf60d767010e23277abbdd2695d8f4',1,'ParaViewModule::ElementInfo']]],
-  ['nodedofnames_5f_19',['nodeDofNames_',['../classGroupOutputModule.html#a1b654e44b18bf8a223bedba5e849ce62',1,'GroupOutputModule']]],
[…]

[…]
