### 2026-08-31 hortulanusT/dynLattice#21 @mohd-afeef-badri
JOSS paper review
@hortulanusT @fmeer

this is part of JOSS review
https://github.com/openjournals/joss-reviews/issues/10410

Thank you for the nicely written JOSS article, I enjoyed reading it. I only have one remark concerning the text:

- For the summary section (which is sort of Abstract of your paper), you choose to not speak at all about your software dynLattice, in my openion it would be best to include couple of phrases about your software in the summary section.
