### 2026-08-31 e7d0196
Explicitly name the software in the Summary section (cf. #21)
