### 2026-09-01 hortulanusT/dynLattice#21 @hortulanusT
Hi @mohd-afeef-badri,

thank you for your kind words, we have added two half-sentences at the end of the summary introducing our software already in #22. As you put this down only as a remark, I took to liberty to already merge the PR and also close this issue.
