### 2026-09-04 #10410 @mbarzegary
@hortulanusT given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR with the editorial corrections
* Work on the author's points of the final checklist I created ☝️

Also, check and fix these two issues in the paper:
* Software-name formatting is inconsistent: JIVE, GMSH, and apptainer are in code font, while MOOSE, FEniCS, deal.II, Kratos, Akantu, and GXBeam are plain. Pick one rule.
* "dynLattice" appears once as plain text ("During these projects, dynLattice was used on a large cluster") while every other occurrence uses [dynLattice]{.sc}.
