### 2026-09-04 a27ec1e
fixed last editorial issues
