### 2026-09-04 cade240
reduced figure width for better spacing
