### 2026-09-04 d5ab52e
increased width a bit again for better spacing
