### 2026-09-04 b4f5daa
bump version to 1.12
