+++ CITATION.cff
@@ -2,7 +2,7 @@ cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 type: software
 title: "dynLattice: A finite element environment for dynamic simulation of beam networks and lattice metamaterials"
-version: "1.11"
+version: "1.12"
 date-released: "2025"
 doi: "10.5281/zenodo.20474969"
 url: "https://github.com/hortulanusT/dynLattice"
