### 2026-09-04 da37374
fix CITATION.cff
