### 2026-09-04 #10410 @hortulanusT
Hi @mbarzegary

thank you (and the two reviewers again!) for your help in this process. I have merged your PR and also fixed the issues you mentioned (I also changed the position of the image slightly, so it should now reader neatly on the first page). I also went through the author's checklist:

- checked the authors + affiliations (ORCIDs are fine)
- I created a release for [v1.12 ](https://github.com/hortulanusT/dynLattice/releases/tag/v1.12) which includes all the latest changes from the Review process.
- the release is archived in Zenodo under https://doi.org/10.5281/zenodo.22310033
- the title and author list match the title and author list in the paper (there is a mismatch in the ordering of my affiliations, from my point of view, this is not a big issue, but let me know if I should make the zenodo order align with the JOSS order to be fully congruent)
- both the github Repo and the Zenodo release show "MIT License"

if there is anything else, that is needed from my side, please let me know!
