### 2026-09-21 ed2a487
Update author format for DHPC entry in bibliography
