+++ paper/paper.bib
@@ -48,7 +48,7 @@
 }
 
 @misc{DHPC2024,
-  author       = {{D}elft {H}igh {P}erformance {C}omputing {C}entre ({DHPC})},
+  author       = {{Delft High Performance Computing Centre (DHPC)}},
   title        = {{D}elft{B}lue {S}upercomputer ({P}hase 2)},
   year         = {2024},
   howpublished = {\url{https://www.tudelft.nl/dhpc/ark:/44463/DelftBluePhase2}},
